ソースファイルの構成を見直し
