+++ elisp/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "elisp"
-version = "0.7.3"
+version = "0.8.0"
 authors = ["hideki kuno <hidekuno@gmail.com>"]
 edition = "2018"
