+++ wasmlisp/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "wasmlisp"
-version = "0.7.3"
+version = "0.8.0"
 authors = ["hidekuno <hidekuno@gmail.com>"]
 edition = "2018"
 description = "A Lisp Interpreter with Web Assembly"
@@ -10,7 +10,7 @@ repository = "https://github.com/hidekuno/rust-elisp"
 crate-type = ["cdylib"]
 
 [dependencies]
-elisp = { path = "../elisp",version = "0.7.3" }
+elisp = { path = "../elisp",version = "0.8.0" }
 wasm-bindgen = "0.2"
 js-sys = "0.3.30"
 wasm-bindgen-test = "0.2"
